+++ src/vector/blas/scale.cpp
@@ -11,7 +11,7 @@ namespace dd_avx{
 			get_isie(x.size(), is, ie);
 			reg alpha_hi = broadcast(alpha.x[0]);
 			reg alpha_lo = broadcast(alpha.x[1]);
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.hi[i]);
 				reg x_lo = load(x.lo[i]);
@@ -36,7 +36,7 @@ namespace dd_avx{
 			get_isie(x.size(), is, ie);
 			reg alpha_hi = broadcast(alpha);
 			reg alpha_lo = regs.zeros;
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.hi[i]);
 				reg x_lo = load(x.lo[i]);
@@ -61,7 +61,7 @@ namespace dd_avx{
 			get_isie(x.size(), is, ie);
 			reg alpha_hi = broadcast(alpha.x[0]);
 			reg alpha_lo = broadcast(alpha.x[1]);
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.data()[i]);
 				reg x_lo = regs.zeros;
@@ -85,7 +85,7 @@ namespace dd_avx{
 			get_isie(x.size(), is, ie);
 			reg alpha_hi = broadcast(alpha);
 			reg alpha_lo = regs.zeros;
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.data()[i]);
 				reg x_lo = regs.zeros;
